Qt: Shuffle some settings around/hide advanced by default
