+++ pcsx2-qt/MainWindow.cpp
@@ -186,8 +186,12 @@ QWidget* MainWindow::getContentParent()
 
 void MainWindow::setupAdditionalUi()
 {
+	const bool show_advanced_settings = QtHost::ShouldShowAdvancedSettings();
+
 	setWindowIcon(QIcon(QStringLiteral("%1/icons/AppIconLarge.png").arg(QtHost::GetResourcesBasePath())));
 	makeIconsMasks(menuBar());
+
+	m_ui.menuDebug->menuAction()->setVisible(show_advanced_settings);
 
 	const bool toolbar_visible = Host::GetBaseBoolSettingValue("UI", "ShowToolbar", false);
 	m_ui.actionViewToolbar->setChecked(toolbar_visible);
@@ -257,7 +261,7 @@ void MainWindow::setupAdditionalUi()
 #ifdef ENABLE_RAINTEGRATION
 	if (Achievements::IsUsingRAIntegration())
 	{
-		QMenu* raMenu = new QMenu(QStringLiteral("RAIntegration"), m_ui.menuDebug);
+		QMenu* raMenu = new QMenu(QStringLiteral("RAIntegration"), m_ui.menu_Tools);
 		connect(raMenu, &QMenu::aboutToShow, this, [this, raMenu]() {
 			raMenu->clear();
 
@@ -282,7 +286,7 @@ void MainWindow::setupAdditionalUi()
 				});
 			}
 		});
-		m_ui.menuDebug->insertMenu(m_ui.actionToggleSoftwareRendering, raMenu);
+		m_ui.menu_Tools->insertMenu(m_ui.menuInput_Recording->menuAction(), raMenu);
 	}
 #endif
 }
@@ -312,7 +316,6 @@ void MainWindow::connectSignals()
 	connect(m_ui.actionGameListSettings, &QAction::triggered, [this]() { doSettings("Game List"); });
 	connect(m_ui.actionEmulationSettings, &QAction::triggered, [this]() { doSettings("Emulation"); });
 	connect(m_ui.actionBIOSSettings, &QAction::triggered, [this]() { doSettings("BIOS"); });
-	connect(m_ui.actionSystemSettings, &QAction::triggered, [this]() { doSettings("System"); });
 	connect(m_ui.actionGraphicsSettings, &QAction::triggered, [this]() { doSettings("Graphics"); });
 	connect(m_ui.actionAudioSettings, &QAction::triggered, [this]() { doSettings("Audio"); });
 	connect(m_ui.actionMemoryCardSettings, &QAction::triggered, [this]() { doSettings("Memory Cards"); });
@@ -372,8 +375,9 @@ void MainWindow::connectSignals()
 	SettingWidgetBinder::BindWidgetToBoolSetting(nullptr, m_ui.actionEnableLogTimestamps, "Logging", "EnableTimestamps", true);
 	SettingWidgetBinder::BindWidgetToBoolSetting(nullptr, m_ui.actionEnableCDVDVerboseReads, "EmuCore", "CdvdVerboseReads", false);
 	SettingWidgetBinder::BindWidgetToBoolSetting(nullptr, m_ui.actionSaveBlockDump, "EmuCore", "CdvdDumpBlocks", false);
+	m_ui.actionShowAdvancedSettings->setChecked(QtHost::ShouldShowAdvancedSettings());
 	connect(m_ui.actionSaveBlockDump, &QAction::toggled, this, &MainWindow::onBlockDumpActionToggled);
-
+	connect(m_ui.actionShowAdvancedSettings, &QAction::toggled, this, &MainWindow::onShowAdvancedSettingsToggled);
 	connect(m_ui.actionSaveGSDump, &QAction::triggered, this, &MainWindow::onSaveGSDumpActionTriggered);
 
 	// Input Recording
@@ -454,10 +458,15 @@ void MainWindow::recreateSettings()
 	QString current_category;
 	if (m_settings_dialog)
 	{
+		const bool was_visible = m_settings_dialog->isVisible();
+
 		current_category = m_settings_dialog->getCategory();
 		m_settings_dialog->hide();
 		m_settings_dialog->deleteLater();
 		m_settings_dialog = nullptr;
+
+		if (!was_visible)
+			return;
 	}
 
 	doSettings(current_category.toUtf8().constData());
@@ -818,6 +827,48 @@ void MainWindow::onBlockDumpActionToggled(bool checked)
 	Host::CommitBaseSettingChanges();
 
 	g_emu_thread->applySettings();
+}
+
+void MainWindow::onShowAdvancedSettingsToggled(bool checked)
+{
+	if (checked && !Host::GetBaseBoolSettingValue("UI", "AdvancedSettingsWarningShown", false))
+	{
+		QCheckBox* cb = new QCheckBox(tr("Do not show again"));
+		QMessageBox mb(this);
+		mb.setWindowTitle(tr("Show Advanced Settings"));
+		mb.setText(
+			tr("Changing advanced settings can have unpredictable effects on games, including graphical glitches, lock-ups, and even corrupted save files. "
+			   "We do not recommend changing advanced settings unless you know what you are doing, and the implications of changing each setting.\n\n"
+			   "The PCSX2 team will not provide any support for configurations that modify these settings, you are on your own.\n\n"
+			   "Are you sure you want to continue?"));
+		mb.setIcon(QMessageBox::Warning);
+		mb.addButton(QMessageBox::Yes);
+		mb.addButton(QMessageBox::No);
+		mb.setDefaultButton(QMessageBox::No);
+		mb.setCheckBox(cb);
+
+		if (mb.exec() == QMessageBox::No)
+		{
+			QSignalBlocker sb(m_ui.actionShowAdvancedSettings);
+			m_ui.actionShowAdvancedSettings->setChecked(false);
+			return;
+		}
+
+		if (cb->isChecked())
+		{
+			Host::SetBaseBoolSettingValue("UI", "AdvancedSettingsWarningShown", true);
+			Host::CommitBaseSettingChanges();
+		}
+	}
+
+	Host::SetBaseBoolSettingValue("UI", "ShowAdvancedSettings", checked);
+	Host::CommitBaseSettingChanges();
+
+	m_ui.menuDebug->menuAction()->setVisible(checked);
+
+	// just recreate the entire settings window, it's easier.
+	if (m_settings_dialog)
+		recreateSettings();
 }
 
 void MainWindow::saveStateToConfig()
@@ -2380,8 +2431,8 @@ void MainWindow::setGameListEntryCoverImage(const GameList::Entry* entry)
 void MainWindow::clearGameListEntryPlayTime(const GameList::Entry* entry)
 {
 	if (QMessageBox::question(this, tr("Confirm Reset"),
-		tr("Are you sure you want to reset the play time for '%1'?\n\nThis action cannot be undone.")
-		.arg(QString::fromStdString(entry->title))) != QMessageBox::Yes)
+			tr("Are you sure you want to reset the play time for '%1'?\n\nThis action cannot be undone.")
+				.arg(QString::fromStdString(entry->title))) != QMessageBox::Yes)
 	{
 		return;
 	}
