+++ pcsx2-qt/Settings/GraphicsSettingsWidget.ui
@@ -322,42 +322,49 @@
        </item>
        <item row="8" column="0" colspan="2">
         <layout class="QGridLayout" name="gridLayout_5">
-         <item row="1" column="1">
-          <widget class="QCheckBox" name="DisableInterlaceOffset">
-           <property name="text">
-            <string>Disable Interlace Offset</string>
-           </property>
-          </widget>
-         </item>
-         <item row="1" column="0">
-          <widget class="QCheckBox" name="vsync">
-           <property name="text">
-            <string>VSync</string>
-           </property>
-          </widget>
-         </item>
-         <item row="0" column="1">
-          <widget class="QCheckBox" name="integerScaling">
-           <property name="text">
-            <string>Integer Upscaling</string>
-           </property>
-          </widget>
-         </item>
-         <item row="2" column="1">
-          <widget class="QCheckBox" name="PCRTCOverscan">
-           <property name="text">
-            <string>Show Overscan</string>
-           </property>
-          </widget>
-         </item>
-         <item row="2" column="0">
+         <item row="3" column="0">
           <widget class="QCheckBox" name="PCRTCOffsets">
            <property name="text">
             <string>Screen Offsets</string>
            </property>
           </widget>
          </item>
+         <item row="1" column="1">
+          <widget class="QCheckBox" name="integerScaling">
+           <property name="text">
+            <string>Integer Upscaling</string>
+           </property>
+          </widget>
+         </item>
+         <item row="2" column="0">
+          <widget class="QCheckBox" name="vsync">
+           <property name="text">
+            <string>VSync</string>
+           </property>
+          </widget>
+         </item>
+         <item row="3" column="1">
+          <widget class="QCheckBox" name="PCRTCOverscan">
+           <property name="text">
+            <string>Show Overscan</string>
+           </property>
+          </widget>
+         </item>
          <item row="0" column="0">
+          <widget class="QCheckBox" name="widescreenPatches">
+           <property name="text">
+            <string>Enable Widescreen Patches</string>
+           </property>
+          </widget>
+         </item>
+         <item row="0" column="1">
+          <widget class="QCheckBox" name="noInterlacingPatches">
+           <property name="text">
+            <string>Enable No-Interlacing Patches</string>
+           </property>
+          </widget>
+         </item>
+         <item row="1" column="0">
           <widget class="QCheckBox" name="PCRTCAntiBlur">
            <property name="text">
             <string>Anti-Blur</string>
@@ -367,12 +374,19 @@
            </property>
           </widget>
          </item>
+         <item row="2" column="1">
+          <widget class="QCheckBox" name="DisableInterlaceOffset">
+           <property name="text">
+            <string>Disable Interlace Offset</string>
+           </property>
+          </widget>
+         </item>
         </layout>
        </item>
        <item row="4" column="0">
-        <widget class="QLabel" name="label_17">
+        <widget class="QLabel" name="label_18">
          <property name="text">
-          <string>Bilinear Filtering</string>
+          <string>Bilinear Filtering:</string>
          </property>
         </widget>
        </item>
@@ -1482,38 +1496,10 @@
               </property>
              </widget>
             </item>
-            <item row="0" column="1">
-             <widget class="QCheckBox" name="osdShowSpeed">
+            <item row="4" column="1">
+             <widget class="QCheckBox" name="osdShowInputs">
               <property name="text">
-               <string>Show Speed</string>
-              </property>
-             </widget>
-            </item>
-            <item row="2" column="0">
-             <widget class="QCheckBox" name="osdShowCPU">
-              <property name="text">
-               <string>Show CPU Usage</string>
-              </property>
-             </widget>
-            </item>
-            <item row="3" column="0">
-             <widget class="QCheckBox" name="osdShowGSStats">
-              <property name="text">
-               <string>Show Statistics</string>
-              </property>
-             </widget>
-            </item>
-            <item row="0" column="0">
-             <widget class="QCheckBox" name="osdShowMessages">
-              <property name="text">
-               <string>Show Notifications</string>
-              </property>
-             </widget>
-            </item>
-            <item row="1" column="0">
-             <widget class="QCheckBox" name="osdShowFPS">
-              <property name="text">
-               <string>Show FPS</string>
+               <string>Show Inputs</string>
               </property>
              </widget>
             </item>
@@ -1531,10 +1517,45 @@
               </property>
              </widget>
             </item>
-            <item row="4" column="1">
-             <widget class="QCheckBox" name="osdShowInputs">
+            <item row="1" column="0">
+             <widget class="QCheckBox" name="osdShowFPS">
               <property name="text">
-               <string>Show Inputs</string>
+               <string>Show FPS</string>
+              </property>
+             </widget>
+            </item>
+            <item row="0" column="0">
+             <widget class="QCheckBox" name="osdShowMessages">
+              <property name="text">
+               <string>Show Notifications</string>
+              </property>
+             </widget>
+            </item>
+            <item row="3" column="0">
+             <widget class="QCheckBox" name="osdShowGSStats">
+              <property name="text">
+               <string>Show Statistics</string>
+              </property>
+             </widget>
+            </item>
+            <item row="2" column="0">
+             <widget class="QCheckBox" name="osdShowCPU">
+              <property name="text">
+               <string>Show CPU Usage</string>
+              </property>
+             </widget>
+            </item>
+            <item row="0" column="1">
+             <widget class="QCheckBox" name="osdShowSpeed">
+              <property name="text">
+               <string>Show Speed</string>
+              </property>
+             </widget>
+            </item>
+            <item row="5" column="0">
+             <widget class="QCheckBox" name="warnAboutUnsafeSettings">
+              <property name="text">
+               <string>Warn About Unsafe Settings</string>
               </property>
              </widget>
             </item>
