+++ pcsx2-qt/Settings/InterfaceSettingsWidget.ui
@@ -74,10 +74,17 @@
         </property>
        </widget>
       </item>
-      <item row="4" column="0">
+      <item row="4" column="1">
        <widget class="QCheckBox" name="discordPresence">
         <property name="text">
          <string>Enable Discord Presence</string>
+        </property>
+       </widget>
+      </item>
+      <item row="4" column="0">
+       <widget class="QCheckBox" name="perGameSettings">
+        <property name="text">
+         <string>Enable Per-Game Settings</string>
         </property>
        </widget>
       </item>
@@ -89,52 +96,48 @@
      <property name="title">
       <string>Game Display</string>
      </property>
-     <layout class="QFormLayout" name="formLayout_3">
-      <item row="0" column="0" colspan="2">
-       <layout class="QGridLayout" name="gridLayout_3">
-        <item row="2" column="0">
-         <widget class="QCheckBox" name="disableWindowResizing">
-          <property name="text">
-           <string>Disable Window Resizing</string>
-          </property>
-         </widget>
-        </item>
-        <item row="0" column="0">
-         <widget class="QCheckBox" name="startFullscreen">
-          <property name="text">
-           <string>Start Fullscreen</string>
-          </property>
-         </widget>
-        </item>
-        <item row="0" column="1">
-         <widget class="QCheckBox" name="doubleClickTogglesFullscreen">
-          <property name="text">
-           <string>Double-Click Toggles Fullscreen</string>
-          </property>
-         </widget>
-        </item>
-        <item row="1" column="0">
-         <widget class="QCheckBox" name="renderToSeparateWindow">
-          <property name="text">
-           <string>Render To Separate Window</string>
-          </property>
-         </widget>
-        </item>
-        <item row="2" column="1">
-         <widget class="QCheckBox" name="hideMouseCursor">
-          <property name="text">
-           <string>Hide Cursor In Fullscreen</string>
-          </property>
-         </widget>
-        </item>
-        <item row="1" column="1">
-         <widget class="QCheckBox" name="hideMainWindow">
-          <property name="text">
-           <string>Hide Main Window When Running</string>
-          </property>
-         </widget>
-        </item>
-       </layout>
+     <layout class="QGridLayout" name="gridLayout_3">
+      <item row="0" column="0">
+       <widget class="QCheckBox" name="startFullscreen">
+        <property name="text">
+         <string>Start Fullscreen</string>
+        </property>
+       </widget>
+      </item>
+      <item row="0" column="1">
+       <widget class="QCheckBox" name="doubleClickTogglesFullscreen">
+        <property name="text">
+         <string>Double-Click Toggles Fullscreen</string>
+        </property>
+       </widget>
+      </item>
+      <item row="1" column="0">
+       <widget class="QCheckBox" name="renderToSeparateWindow">
+        <property name="text">
+         <string>Render To Separate Window</string>
+        </property>
+       </widget>
+      </item>
+      <item row="1" column="1">
+       <widget class="QCheckBox" name="hideMainWindow">
+        <property name="text">
+         <string>Hide Main Window When Running</string>
+        </property>
+       </widget>
+      </item>
+      <item row="2" column="0">
+       <widget class="QCheckBox" name="disableWindowResizing">
+        <property name="text">
+         <string>Disable Window Resizing</string>
+        </property>
+       </widget>
+      </item>
+      <item row="2" column="1">
+       <widget class="QCheckBox" name="hideMouseCursor">
+        <property name="text">
+         <string>Hide Cursor In Fullscreen</string>
+        </property>
+       </widget>
       </item>
      </layout>
     </widget>
