+++ pcsx2-qt/Settings/SettingsDialog.h
@@ -33,8 +33,6 @@ class InterfaceSettingsWidget;
 class GameListSettingsWidget;
 class EmulationSettingsWidget;
 class BIOSSettingsWidget;
-class SystemSettingsWidget;
-class AdvancedSystemSettingsWidget;
 class GameFixSettingsWidget;
 class GraphicsSettingsWidget;
 class AudioSettingsWidget;
@@ -42,6 +40,7 @@ class MemoryCardSettingsWidget;
 class FolderSettingsWidget;
 class DEV9SettingsWidget;
 class AchievementSettingsWidget;
+class AdvancedSettingsWidget;
 
 class SettingsDialog final : public QDialog
 {
@@ -61,8 +60,6 @@ public:
 	__fi GameListSettingsWidget* getGameListSettingsWidget() const { return m_game_list_settings; }
 	__fi BIOSSettingsWidget* getBIOSSettingsWidget() const { return m_bios_settings; }
 	__fi EmulationSettingsWidget* getEmulationSettingsWidget() const { return m_emulation_settings; }
-	__fi SystemSettingsWidget* getSystemSettingsWidget() const { return m_system_settings; }
-	__fi AdvancedSystemSettingsWidget* getAdvancedSystemSettingsWidget() const { return m_advanced_system_settings; }
 	__fi GameFixSettingsWidget* getGameFixSettingsWidget() const { return m_game_fix_settings_widget; }
 	__fi GraphicsSettingsWidget* getGraphicsSettingsWidget() const { return m_graphics_settings; }
 	__fi AudioSettingsWidget* getAudioSettingsWidget() const { return m_audio_settings; }
@@ -70,6 +67,7 @@ public:
 	__fi FolderSettingsWidget* getFolderSettingsWidget() const { return m_folder_settings; }
 	__fi DEV9SettingsWidget* getDEV9SettingsWidget() const { return m_dev9_settings; }
 	__fi AchievementSettingsWidget* getAchievementSettingsWidget() const { return m_achievement_settings; }
+	__fi AdvancedSettingsWidget* getAdvancedSettingsWidget() const { return m_advanced_settings; }
 
 	void registerWidgetHelp(QObject* object, QString title, QString recommended_value, QString text);
 	bool eventFilter(QObject* object, QEvent* event) override;
@@ -121,8 +119,6 @@ private:
 	GameListSettingsWidget* m_game_list_settings = nullptr;
 	BIOSSettingsWidget* m_bios_settings = nullptr;
 	EmulationSettingsWidget* m_emulation_settings = nullptr;
-	SystemSettingsWidget* m_system_settings = nullptr;
-	AdvancedSystemSettingsWidget* m_advanced_system_settings = nullptr;
 	GameFixSettingsWidget* m_game_fix_settings_widget = nullptr;
 	GraphicsSettingsWidget* m_graphics_settings = nullptr;
 	AudioSettingsWidget* m_audio_settings = nullptr;
@@ -130,6 +126,7 @@ private:
 	FolderSettingsWidget* m_folder_settings = nullptr;
 	DEV9SettingsWidget* m_dev9_settings = nullptr;
 	AchievementSettingsWidget* m_achievement_settings = nullptr;
+	AdvancedSettingsWidget* m_advanced_settings = nullptr;
 
 	std::array<QString, MAX_SETTINGS_WIDGETS> m_category_help_text;
