+++ pcsx2-qt/Settings/SettingsDialog.ui
@@ -33,13 +33,13 @@
      </property>
      <property name="minimumSize">
       <size>
-       <width>150</width>
+       <width>160</width>
        <height>0</height>
       </size>
      </property>
      <property name="maximumSize">
       <size>
-       <width>150</width>
+       <width>160</width>
        <height>16777215</height>
       </size>
      </property>
